Create packages to organize the commands and started with testing.
