+++ pravega-cli/src/main/java/io/pravega/tools/pravegacli/commands/bookkeeper/BookKeeperEnableCommand.java
