+++ pravega-cli/src/main/java/io/pravega/tools/pravegacli/commands/config/ConfigCommand.java
@@ -7,7 +7,10 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.tools.pravegacli.commands;
+package io.pravega.tools.pravegacli.commands.config;
+
+import io.pravega.tools.pravegacli.commands.Command;
+import io.pravega.tools.pravegacli.commands.CommandArgs;
 
 /**
  * Base for all Config-related commands.
