+++ pravega-cli/src/main/java/io/pravega/tools/pravegacli/commands/controller/ControllerListScopesCommand.java
@@ -1,0 +1,4 @@
+package io.pravega.tools.pravegacli.commands;
+
+public class ControllerListScopesCommand {
+}
