+++ pravega-cli/src/test/java/io/pravega/tools/pravegacli/BookkeeperCommandsTest.java
@@ -1,0 +1,4 @@
+package io.pravega.tools.pravegacli;
+
+public class BookkeeperCommandsTest {
+}
