+++ pravega-cli/src/test/java/io/pravega/tools/pravegacli/ControllerCommandsTest.java
@@ -1,0 +1,4 @@
+package io.pravega.tools.pravegacli;
+
+public class ControllerCommandsTest {
+}
